Major namespace changes to refactor out opentok name
