+++ tests/VonageVideoTest/Util/responses/signal-failure-invalid-token.json
